Extract PartComponentLink and show similar kanji in popovers.
Move shared component-link helpers into reusable modules so visually similar kanji can open a PartComponentLink popover instead of the full breakdown.
