+++ src/components/dependent/routing/global-links.tsx
@@ -4,6 +4,31 @@ import { useKanjiFromUrl, useUrlLocation } from "./routing-hooks";
 import { Link } from "./router-adapter";
 import { radicalFalseFriends } from "@/lib/radicals";
 
+export const ComponentLink = ({
+  component,
+  keyword,
+  title,
+  type
+}: {
+  component: string;
+  keyword: string;
+  title?: string;
+  type: 'kanji' | 'radical' | 'unknown'
+}) => {
+
+  return (
+    <div className="flex flex-col text-center w-fit ">
+      {type === 'kanji' ?
+        <GlobalKanjiLink kanji={component} keyword={keyword} /> : type === "radical" ?
+          <GlobalRadicalLink radical={component} keyword={keyword} /> : <FakeComponentLink radical={component} keyword={keyword} />
+      }
+      {title && <div className="text-[10px] uppercase opacity-70">{title}</div>}
+    </div>
+  )
+
+
+};
+
 type FontSize = 'text-xl' | 'text-2xl' | 'text-3xl' | 'text-4xl' | 'text-5xl' | 'text-6xl' | 'text-7xl' | 'text-8xl' | 'text-9xl' | 'text-10xl';
 
 const cnJPCard = "flex flex-col m-1 p-1 text-xl rounded-md";
